Initial commit.
This is the first pass at the system. It stores account information in
Postgres, with the more sensitive details stored in Vault. It downloads
all transactions from OFX, and stores them in Postgres, as well.

NOTE: at the moment, sensitive information (credit card numbers, routing
and account numbers, etc.) will be written to Postgres. A plan is in
place to fix this, but it involves data migration and no automated
tooling will be provided to do that migration.

There's also a `similar` package that implements shingling and the
jaccard coefficient, which can be used to group transactions by their
names, as group.go does.

Right now there are basically no interfaces to this. the `cmd` folder
has a few packages that all build binaries:

* budget-groups prints out the groups of transactions, as a test of that
  logic.
* budget-migrations applies SQL migrations to the database.
* budget-import fetches transactions using OFX and imports them into the
  database.

There are also very, very few tests in this project at the moment. Life
goes on. Eventually, there will be. But for the moment, this is enough
functionality and progress that I don't want to lose it, and so getting
it off my computer is desirable.