Drop sensitive fields, only sync some accounts.
Some accounts will be banks that don't support OFX, but we still want to
be able to manually enter transactions for them, or load in the files
they provide and associate those transactions with them. So now Accounts
support a Sync flag and `budget-import` will only attempt to fetch
transactions for accounts over OFX if the Account has the Sync flag set
to `true`.

We also completed the migration from storing (our) sensitive information
in Postgres, by dropping the columns and fixing our unique constraint to
use the account_id instead of the combination of source parameters. Note
that all data migration (filling in the correct account_id, setting sync
to true on accounts) was done manually, not through the
`budget-migrations` binary.