Recurring -> Groups, add Trends, small fixes.
Rename recurring to groups, as its only responsibility now is to track
groups of similar transactions, regardless of whether they're recurring
or not.

Add trends, which track groups of groups (and transactions) and forecast
them into the future, and is how we model recurring transactions now.

Sort transactions by date posted, account, and amount now when listing
transactions.

Add a command to run the server to budgetctl.